create a node that publishes simulated sensors values
